baseline-langchain/pyproject.toml
[project]
name = "janus-baseline-langchain"
version = "0.1.0"
description = "Janus Baseline Competitor - LangChain implementation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "langchain>=0.1.0,<0.2.0",
    "langchain-openai>=0.0.5,<0.1.0",
    "langchain-community>=0.0.20,<0.1.0",
    "langchain-experimental>=0.0.50,<0.1.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "structlog>=24.1.0",
    "tavily-python>=0.3.0",
]

[project.scripts]
janus-baseline-langchain = "janus_baseline_langchain.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "respx>=0.20.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"
